test(SecurityUtils): adding test util to setup temp folder with security database setup and tear down.
